requirements.txt
# Core Framework - relaxed versions for compatibility
langgraph>=0.0.1
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-openai>=0.1.0
autogen-agentchat>=0.4.2

# LLM and AI
anthropic>=0.20.0
openai>=1.0.0
transformers>=4.30.0

# Vector Database
chromadb>=0.3.0

# NLP
spacy>=3.5.0
nltk>=3.8.0

# Document Generation
python-docx>=1.0.0
reportlab>=4.0.0
pypdf2>=3.0.0
markdown>=3.4.0
pymupdf>=1.23.0
pdfminer.six>=20221105

# Visualization
plotly>=5.0.0
seaborn>=0.12.0
matplotlib>=3.6.0

# API and Web
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.24.0

# Database
sqlalchemy>=2.0.0
alembic>=1.10.0

# Academic APIs
semanticscholar>=0.1.0
arxiv>=1.4.0
requests>=2.30.0
# paperswithcode-client - DISABLED: has incompatible old dependencies

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
jinja2>=3.1.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
certifi>=2023.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pylint>=2.17.0
isort>=5.12.0

# Logging
loguru>=0.7.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Development Tools
ipython>=8.0.0
jupyter>=1.0.0
